refactor(chat): extract ReadReceipt and FileAttachment from MessageDisplay
Stage 4a of the MessageDisplay decomposition. Two verbatim component
extractions, no behavior change:

- ReadReceipt.jsx  ← renderReadReceipt (tooltip resolution stays in parent)
- FileAttachment.jsx ← renderFileAttachment
- getFileIcon moved to messageDisplayHelpers.js (also used by the
  reply-preview block, so kept shared rather than colocated)

Removes now-unused imports (CheckCheck, File, FileSpreadsheet,
formatFileSize). MessageDisplay 4320 → 4194 lines. ESLint + build green.
